refactor: 계약서 분석 공통 부분 메인 컴포넌트에서 관리
