+++ src/features/contract-analysis/main/pages/ContractAnalysisMain.jsx
@@ -1,18 +1,13 @@
-import { BackButton } from '@/components/BackButton/BackButton'
 import { PrimaryButton } from '@/components/PrimaryButton/PrimaryButton'
 
 import { ContractAnalysisHeader } from '../components/ContractAnalysisHeader'
 import { ContractAnalysisStepList } from '../components/ContractAnalysisStepList'
 
-import styles from './ContractAnalysis.module.css'
+import styles from './ContractAnalysisMain.module.css'
 
-export const ContractAnalysis = () => {
+export const ContractAnalysisMain = () => {
   return (
-    <div className={styles['container']}>
-      <nav className={styles['back-button']}>
-        <BackButton />
-      </nav>
-
+    <>
       <section className={styles['header']}>
         <ContractAnalysisHeader />
       </section>
@@ -24,6 +19,6 @@ export const ContractAnalysis = () => {
       <div className={styles['start-button']}>
         <PrimaryButton size='lg' label='시작하기' />
       </div>
-    </div>
+    </>
   )
 }
