+++ src/features/contract-analysis/main/pages/ContractAnalysisMain.module.css
@@ -1,13 +1,3 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-.back-button {
-  margin: 48px 16px 24px;
-}
-
 .header {
   margin: 0 36px 30px;
 }
