+++ src/features/contract-analysis/pages/ContractAnalysis.module.css
@@ -1,0 +1,9 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+.back-button {
+  margin: 48px 16px 24px;
+}
